feat: Allow temp paths in pathmapping destination; cleanup windows transformers dir structure
