+++ transformer/classes/dockerfile/windows/consoleappdockerfilegenerator.go
@@ -14,13 +14,17 @@
  *  limitations under the License.
  */
 
-package dotnet
+package windows
 
 import (
 	"encoding/xml"
+	"fmt"
 	"io/ioutil"
+	"net"
+	"net/url"
 	"os"
 	"path/filepath"
+	"strconv"
 	"strings"
 
 	"github.com/konveyor/move2kube/common"
@@ -32,43 +36,133 @@ import (
 	"github.com/sirupsen/logrus"
 )
 
-// SilverLightTemplateConfig implements SilverLight config interface
-type SilverLightTemplateConfig struct {
-	Ports   []int32
-	AppName string
+const (
+	// AppConfigFilePathListType points to the go.mod file path
+	AppConfigFilePathListType transformertypes.PathType = "AppConfigFilePathList"
+
+	// AppCfgFile is file name of App.Config file in dotnet projects
+	AppCfgFile = "App.config"
+)
+
+// ConsoleTemplateConfig implements .Net Console config interface
+type ConsoleTemplateConfig struct {
+	Ports            []int32
+	AppName          string
+	BaseImageVersion string
 }
 
-// WinSilverLightWebAppDockerfileGenerator implements the Transformer interface
-type WinSilverLightWebAppDockerfileGenerator struct {
+// WinConsoleAppDockerfileGenerator implements the Transformer interface
+type WinConsoleAppDockerfileGenerator struct {
 	Config transformertypes.Transformer
 	Env    *environment.Environment
 }
 
 // Init Initializes the transformer
-func (t *WinSilverLightWebAppDockerfileGenerator) Init(tc transformertypes.Transformer, env *environment.Environment) (err error) {
+func (t *WinConsoleAppDockerfileGenerator) Init(tc transformertypes.Transformer, env *environment.Environment) (err error) {
 	t.Config = tc
 	t.Env = env
 	return nil
 }
 
 // GetConfig returns the transformer config
-func (t *WinSilverLightWebAppDockerfileGenerator) GetConfig() (transformertypes.Transformer, *environment.Environment) {
+func (t *WinConsoleAppDockerfileGenerator) GetConfig() (transformertypes.Transformer, *environment.Environment) {
 	return t.Config, t.Env
 }
 
 // BaseDirectoryDetect runs detect in base directory
-func (t *WinSilverLightWebAppDockerfileGenerator) BaseDirectoryDetect(dir string) (namedServices map[string]transformertypes.ServicePlan, unnamedServices []transformertypes.TransformerPlan, err error) {
+func (t *WinConsoleAppDockerfileGenerator) BaseDirectoryDetect(dir string) (namedServices map[string]transformertypes.ServicePlan, unnamedServices []transformertypes.TransformerPlan, err error) {
 	return nil, nil, nil
 }
 
+// parseAppConfig parses the application config
+func (t *WinConsoleAppDockerfileGenerator) parseAppConfigForPort(AppCfgFilePath string) ([]int32, error) {
+	appConfigFile, err := os.Open(AppCfgFilePath)
+	if err != nil {
+		return nil, fmt.Errorf("could not open the App.config file: %s", err)
+	}
+
+	defer appConfigFile.Close()
+
+	byteValue, _ := ioutil.ReadAll(appConfigFile)
+	appCfg := dotnet.AppConfig{}
+	xml.Unmarshal(byteValue, &appCfg)
+	if err != nil {
+		return nil, fmt.Errorf("could not parse the App.config file: %s", err)
+	}
+
+	ports := make([]int32, 0)
+	for _, addKey := range appCfg.AppCfgSettings.AddList {
+		parsedURL, err := url.ParseRequestURI(addKey.Value)
+		if err != nil {
+			logrus.Errorf("Could not parse URI: %s", err)
+			continue
+		}
+
+		if parsedURL.Scheme == "" || parsedURL.Host == "" {
+			logrus.Warnf("Scheme or host is empty in URI")
+			continue
+		}
+
+		_, port, err := net.SplitHostPort(parsedURL.Host)
+		if err != nil {
+			logrus.Errorf("Could not extract port from URI: %s", err)
+			continue
+		}
+
+		portAsInt, err := strconv.ParseInt(port, 10, 32)
+		if err != nil {
+			logrus.Errorf("Could not process port from URI: %s", err)
+			continue
+		}
+
+		ports = append(ports, int32(portAsInt))
+	}
+
+	if len(appCfg.Model.Services.ServiceList) == 0 {
+		return ports, nil
+	}
+
+	for _, svc := range appCfg.Model.Services.ServiceList {
+		for _, addKey := range svc.Host.BaseAddresses.AddList {
+			parsedURL, err := url.ParseRequestURI(addKey.BaseAddress)
+			if err != nil {
+				logrus.Errorf("Could not parse URI: %s", err)
+				continue
+			}
+
+			if parsedURL.Scheme == "" || parsedURL.Host == "" {
+				logrus.Warnf("Scheme or host is empty in URI")
+				continue
+			}
+
+			_, port, err := net.SplitHostPort(parsedURL.Host)
+			if err != nil {
+				logrus.Errorf("Could not extract port from URI: %s", err)
+				continue
+			}
+
+			portAsInt, err := strconv.ParseInt(port, 10, 32)
+			if err != nil {
+				logrus.Errorf("Could not process port from URI: %s", err)
+				continue
+			}
+
+			ports = append(ports, int32(portAsInt))
+		}
+	}
+
+	return ports, nil
+}
+
 // DirectoryDetect runs detect in each sub directory
-func (t *WinSilverLightWebAppDockerfileGenerator) DirectoryDetect(dir string) (namedServices map[string]transformertypes.ServicePlan, unnamedServices []transformertypes.TransformerPlan, err error) {
+func (t *WinConsoleAppDockerfileGenerator) DirectoryDetect(dir string) (namedServices map[string]transformertypes.ServicePlan, unnamedServices []transformertypes.TransformerPlan, err error) {
 	dirEntries, err := os.ReadDir(dir)
 	if err != nil {
 		logrus.Errorf("Error while trying to read directory: %s", err)
 		return nil, nil, err
 	}
 	appName := ""
+	appConfigList := make([]string, 0)
 	for _, de := range dirEntries {
 		if filepath.Ext(de.Name()) != dotnet.CsSln {
 			continue
@@ -111,18 +205,15 @@ func (t *WinSilverLightWebAppDockerfileGenerator) DirectoryDetect(dir string) (n
 				logrus.Errorf("%s", err)
 				continue
 			}
-			if !isWebProj {
+			if isWebProj {
 				continue
 			}
 
-			isSLProj, err := isSilverlight(configuration)
-			if err != nil {
-				logrus.Errorf("%s", err)
+			appCfgFilePath := filepath.Join(dir, filepath.Dir(csPath), AppCfgFile)
+			if _, err = os.Stat(appCfgFilePath); os.IsNotExist(err) {
 				continue
 			}
-			if !isSLProj {
-				continue
-			}
+			appConfigList = append(appConfigList, appCfgFilePath)
 
 			appName = strings.TrimSuffix(filepath.Base(de.Name()), filepath.Ext(de.Name()))
 		}
@@ -142,6 +233,7 @@ func (t *WinSilverLightWebAppDockerfileGenerator) DirectoryDetect(dir string) (n
 			BaseArtifactTypes: []transformertypes.ArtifactType{artifacts.ContainerBuildArtifactType},
 			Paths: map[string][]string{
 				artifacts.ProjectPathPathType: {dir},
+				AppConfigFilePathListType:     appConfigList,
 			},
 		}},
 	}
@@ -149,7 +241,7 @@ func (t *WinSilverLightWebAppDockerfileGenerator) DirectoryDetect(dir string) (n
 }
 
 // Transform transforms the artifacts
-func (t *WinSilverLightWebAppDockerfileGenerator) Transform(newArtifacts []transformertypes.Artifact, oldArtifacts []transformertypes.Artifact) ([]transformertypes.PathMapping, []transformertypes.Artifact, error) {
+func (t *WinConsoleAppDockerfileGenerator) Transform(newArtifacts []transformertypes.Artifact, oldArtifacts []transformertypes.Artifact) ([]transformertypes.PathMapping, []transformertypes.Artifact, error) {
 	pathMappings := []transformertypes.PathMapping{}
 	artifactsCreated := []transformertypes.Artifact{}
 	for _, a := range newArtifacts {
@@ -173,11 +265,22 @@ func (t *WinSilverLightWebAppDockerfileGenerator) Transform(newArtifacts []trans
 		}
 
 		var detectedPorts []int32
-		detectedPorts = append(detectedPorts, 8080) //TODO: Write parser to parse and identify port
+		for _, appConfigFilePath := range a.Paths[AppConfigFilePathListType] {
+			portList, err := t.parseAppConfigForPort(appConfigFilePath)
+			if err != nil {
+				logrus.Errorf("%s", err)
+				continue
+			}
+			if portList != nil {
+				detectedPorts = append(detectedPorts, portList...)
+			}
+		}
+
 		detectedPorts = commonqa.GetPortsForService(detectedPorts, a.Name)
-		var silverLightConfig SilverLightTemplateConfig
-		silverLightConfig.AppName = a.Name
-		silverLightConfig.Ports = detectedPorts
+		var consoleConfig ConsoleTemplateConfig
+		consoleConfig.AppName = a.Name
+		consoleConfig.Ports = detectedPorts
+		consoleConfig.BaseImageVersion = dotnet.DefaultBaseImageVersion
 
 		if sImageName.ImageName == "" {
 			sImageName.ImageName = common.MakeStringContainerImageNameCompliant(sConfig.ServiceName)
@@ -189,7 +292,7 @@ func (t *WinSilverLightWebAppDockerfileGenerator) Transform(newArtifacts []trans
 			Type:           transformertypes.TemplatePathMappingType,
 			SrcPath:        filepath.Join(t.Env.Context, t.Config.Spec.TemplatesDir),
 			DestPath:       filepath.Join(common.DefaultSourceDir, relSrcPath),
-			TemplateConfig: silverLightConfig,
+			TemplateConfig: consoleConfig,
 		})
 		paths := a.Paths
 		paths[artifacts.DockerfilePathType] = []string{filepath.Join(common.DefaultSourceDir, relSrcPath, "Dockerfile")}
